requirements.txt
pillow == 9.3.0
flask == 2.2.2